Change app endpoint rendered component; Refactor NoteListPage to contain only its own functionality
