+++ frontend/src/features/app/note-list/NoteListPage.tsx
@@ -1,151 +1,15 @@
-import { useMemo, useState } from 'react';
-import AppNavbarComponent from '../../../common-component/AppNavbarComponent';
-import { NoteType } from '../typedefs/NoteType';
-import { TagCollectionType } from '../typedefs/TagCollectionType';
-import NoteListTabComponent from './components/NoteListTabComponent';
-import NoteTabComponent from './components/NoteTabComponent';
 import SearchbarComponent from './components/SearchbarComponent';
-import TagCardComponent from './components/TagCardComponent';
-import TagListTabComponent from './components/TagListTabComponent';
-import NoteListPageTabEnum from './typedefs/NoteListPageTabEnum';
-import NoteDataMapper from './utils/NoteDataMapper';
 
-const dummyData: NoteType[] = [
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#E8CBA9', text_color: '#615547' },
-      { name: 'Project', body_color: '#E5CEE0', text_color: '#4B4049' },
-    ],
-    timestamp: 1653978467000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Test', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1653892064000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Project', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1653978464000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Test', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1653792064000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Project', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1653988464000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Test', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1653892024000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Project', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1693978464000,
-  },
-  {
-    title: 'Covalent Bonding',
-    body: 'A covalent bond is a chemical bond that involves the sharing of electron pairs between atoms.These electron pairs are known as shared pairs or bonding pairs, and the stable balance of attractive an...',
-    tag: [
-      { name: 'Chemistry', body_color: '#000000', text_color: '#FFFFFF' },
-      { name: 'Test', body_color: '#000000', text_color: '#FFFFFF' },
-    ],
-    timestamp: 1553892064000,
-  },
-];
+type Params = {
+  noteComponents: React.ReactNode[],
+};
 
-const dummyTags: TagCollectionType[] = [
-  {
-    name: 'Project',
-    body_color: '#4F4F4F',
-    text_color: '#FFFFFF',
-    total: 12,
-  },
-  {
-    name: 'Chemistry',
-    body_color: '#E8CBA9',
-    text_color: '#615547',
-    total: 12,
-  },
-  {
-    name: 'Physics',
-    body_color: '#E5CEE0',
-    text_color: '#4B4049',
-    total: 12,
-  },
-  {
-    name: 'Algorithm',
-    body_color: '#8EA47E',
-    text_color: '#F8FFEF',
-    total: 12,
-  },
-];
-
-export default function NoteListPage() {
-  const [currentTab, setCurrentTab] = useState<NoteListPageTabEnum>(NoteListPageTabEnum.NoteList);
-
-  const sortedData = dummyData.sort((a, b) => b.timestamp - a.timestamp);
-  const generatedComponent = useMemo(() => NoteDataMapper(sortedData), [sortedData]);
-  const generatedTags = useMemo(() => dummyTags.map(
-    (element) => <TagCardComponent tag={element} />,
-  ), []);
-
-  const tabComponent = useMemo(() => {
-    if (currentTab === NoteListPageTabEnum.NoteList) {
-      return <NoteListTabComponent noteComponents={generatedComponent} />;
-    }
-
-    if (currentTab === NoteListPageTabEnum.TagList) {
-      return <TagListTabComponent tagComponents={generatedTags} />;
-    }
-
-    return <NoteTabComponent />;
-  }, [currentTab, generatedComponent, generatedTags]);
-
+export default function NoteListPage({ noteComponents }: Params) {
   return (
-    <div id="page-container" className="flex flex-row w-screen h-screen">
-      <AppNavbarComponent
-        currentTab={currentTab}
-        onNewNoteClick={() => console.log('implement')}
-        onNoteListTabClick={() => setCurrentTab(NoteListPageTabEnum.NoteList)}
-        onTagListTabClick={() => setCurrentTab(NoteListPageTabEnum.TagList)}
-        onNoteTabClick={() => setCurrentTab(NoteListPageTabEnum.Note)}
-        onUserClick={() => console.log('implement')}
-      />
-      <div id="app-content" className="flex flex-col w-full h-full">
-        <SearchbarComponent />
-        {tabComponent}
+    <div className="w-full h-full bg-white">
+      <SearchbarComponent />
+      <div className="flex overflow-y-scroll flex-col py-8 px-12 w-full h-full">
+        {noteComponents}
       </div>
     </div>
   );
