+++ frontend/src/features/app/note-list/components/NoteListTabComponent.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import SearchbarComponent from './SearchbarComponent';
 
 type Params = {
   noteComponents: React.ReactNode[],
@@ -6,8 +7,11 @@ type Params = {
 
 export default function NoteListTabComponent({ noteComponents }: Params) {
   return (
-    <div className="flex overflow-y-scroll flex-col py-8 px-12 w-full">
-      {noteComponents}
+    <div className="bg-white">
+      <SearchbarComponent />
+      <div className="flex overflow-y-scroll flex-col py-8 px-12 w-full">
+        {noteComponents}
+      </div>
     </div>
   );
 }
