+++ frontend/src/features/app/note-page/NotePage.tsx
@@ -1,0 +1,3 @@
+export default function NotePage() {
+  return <div>Reserved</div>;
+}
